Refactor & simplify comparison
